orginze funcions
added new round dealing
