+++ frontend/gameplan.js
@@ -1,44 +1,42 @@
-loadGameSetup();
-// players creation
-const players = [{ player1 }, { player2 }, { player3 }, { player4 }];
-createGame(gameControl) {
-    while (gameControl.players.length > 1) {
-        round(players);
-        gameControlUpdate(players)
-        //   -optional removePlayers();
-        scoreUpdate(players);
-        // 1 function that hids the welcome page
-        // 2 method in player that reset the score and sums it in sum property
-        // 3 build scoreUpDate to be array of objects each object resamble a round in the game
-        // [{player1: score, player2}]
-        showScore();
-    }
+// loadGameSetup();
+// // players creation
+// const players = [{ player1 }, { player2 }, { player3 }, { player4 }];
+// createGame(gameControl) {
+//     while (gameControl.players.length > 1) {
+//         round(players);
+//         gameControlUpdate(players)
+//         //   -optional removePlayers();
+//         scoreUpdate(players);
+//         // 1 build scoreUpDate to be array of objects each object resamble a round in the game
+//         // [{player1: score, player2}]
+//         showScore();
+//     }
 
-    // }; // Deck => playerDeck TableDeck', 'PileDeck'for each player
-    async function round(players) {
-        const currentPlayer = player[0]
-        while (!Yaniv) {
-            await turn(currentPlayer);
-            passTurn();
-        }
+//     // }; // Deck => playerDeck TableDeck', 'PileDeck'for each player
+//     async function round(players) {
+//         const currentPlayer = player[0]
+//         while (!Yaniv) {
+//             await turn(currentPlayer);
+//             passTurn();
+//         }
 
-        scoreCount(players);
+//         scoreCount(players);
 
-    };
-    startTurn(player);
-    callYaniv();
-    chooseCard();
+//     };
+//     startTurn(player);
+//     callYaniv();
+//     chooseCard();
 
-    function trun(player) {
-        player.throwCards(cards);
-        player.drawTableCard(deck.drawCard())
-    }
-    // function removePlayer(player) {
+//     function trun(player) {
+//         player.throwCards(cards);
+//         player.drawTableCard(deck.drawCard())
+//     }
+//     // function removePlayer(player) {
 
-}
+// }
 
-scoreCount(players) {
-    for (const player in players) {
-        players.score = sum(players.cards);
-    }
-}
+// scoreCount(players) {
+//     for (const player in players) {
+//         players.score = sum(players.cards);
+//     }
+// }
